# docs/swagger.yaml
definitions:
  handler.CreateOpeningRequest:
    properties:
      company:
        type: string
      link:
        type: string
      location:
        type: string
      remote:
        type: boolean
      role:
        type: string
      salary:
        type: integer
    type: object
  handler.ErrorResponse:
    properties:
      errorCode:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
paths:
  /opening:
    delete:
      consumes:
      - application/json
      description: Delete a new job opening
      parameters:
      - description: Opening identification
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Delete openings
      tags:
      - Openings
    get:
      consumes:
      - application/json
      description: Show a job opening
      parameters:
      - description: Opening identification
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Show openings
      tags:
      - Openings
    post:
      consumes:
      - application/json
      description: Create a new job opening
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateOpeningRequest'
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create openings
      tags:
      - Openings
swagger: "2.0"